* various tidyups to the makefiles - now OPCODE builds from the top   level makefile. * fixed up the DLL defs files, now there are with- and without-OPCODE   versions that should both work. * use 'size_t' instead of 'int' for the allocator function arguments,   and in various other places. * fixed a few warnings (loss of precision, implicit cast to bool). * removed the debugging allocator from memory.cpp, since it is very old   and there are much better alternatives now.

+++ include/ode/common.h
@@ -159,7 +159,7 @@ typedef dReal dQuaternion[4];
  */
 
 #define dALLOCA16(n) \
-  ((char*)dEFFICIENT_SIZE(((int)(alloca((n)+(EFFICIENT_ALIGNMENT-1))))))
+  ((char*)dEFFICIENT_SIZE(((size_t)(alloca((n)+(EFFICIENT_ALIGNMENT-1))))))
 
 
 /* internal object types (all prefixed with `dx') */
